Refactor to REST conventions
* Refactor endpoints to use REST conventions for naming and for POST / GET methods
* Separate out endpoints to individual files per tag.
* Update tests to avoid potential data conflict.
