+++ MyWhiskyShelf.IntegrationTests/Fixtures/MyWhiskyShelfFixtureEmptyDb.cs
@@ -3,10 +3,4 @@ using JetBrains.Annotations;
 namespace MyWhiskyShelf.IntegrationTests.Fixtures;
 
 [UsedImplicitly]
-public class MyWhiskyShelfBaseFixtureEmptyDb : MyWhiskyShelfBaseFixture
-{
-    public MyWhiskyShelfBaseFixtureEmptyDb()
-    {
-        UseDataSeeding = false;
-    }
-}
+public class MyWhiskyShelfBaseFixtureEmptyDb : MyWhiskyShelfBaseFixture;
